+++ lib/imap/backup/account/connection.rb
@@ -27,6 +27,10 @@ module Imap::Backup
         Account::Connection::BackupFolders.new(client: client, account: account).run
     end
 
+    def namespaces
+      client.namespace
+    end
+
     def run_backup(refresh: false)
       Logger.logger.info "Running backup of account: #{account.username}"
       # start the connection so we get logging messages in the right order