+++ lib/imap/backup/cli/remote.rb
@@ -21,6 +21,30 @@ module Imap::Backup
       end
     end
 
+    desc "namespaces EMAIL", "List account namespaces"
+    long_desc <<~DESC
+      Lists namespaces defined for an email account.
+
+      This command is useful in deciding the parameters for
+      the `imap-backup migrate` and `imap-backup mirror` commands.
+    DESC
+    config_option
+    format_option
+    quiet_option
+    verbose_option
+    def namespaces(email)
+      Imap::Backup::Logger.setup_logging options
+      config = load_config(**options)
+      connection = connection(config, email)
+      namespaces = connection.namespaces
+      case options[:format]
+      when "json"
+        json_format_namespaces namespaces
+      else
+        list_namespaces namespaces
+      end
+    end
+
     no_commands do
       def names(email)
         config = load_config(**options)
@@ -40,6 +64,43 @@ module Imap::Backup
         names.each do |name|
           Kernel.puts %("#{name}")
         end
+      end
+
+      def json_format_namespaces(namespaces)
+        list = {
+          personal: namespace_info(namespaces.personal.first),
+          other: namespace_info(namespaces.other.first),
+          shared: namespace_info(namespaces.shared.first)
+        }
+        Kernel.puts list.to_json
+      end
+
+      def list_namespaces(namespaces)
+        Kernel.puts format(
+          "%-10<name>s %-10<prefix>s %<delim>s",
+          {name: "Name", prefix: "Prefix", delim: "Delimiter"}
+        )
+        list_namespace namespaces, :personal
+        list_namespace namespaces, :other
+        list_namespace namespaces, :shared
+      end
+
+      def list_namespace(namespaces, name)
+        info = namespace_info(namespaces.send(name).first, quote: true)
+        if info
+          Kernel.puts format("%-10<name>s %-10<prefix>s %<delim>s", name: name, **info)
+        else
+          Kernel.puts format("%-10<name>s (Not defined)", name: name)
+        end
+      end
+
+      def namespace_info(namespace, quote: false)
+        return nil if !namespace
+
+        {
+          prefix: quote ? namespace.prefix.to_json : namespace.prefix,
+          delim: quote ? namespace.delim.to_json : namespace.delim
+        }
       end
     end
   end