+++ lib/imap/backup/client/default.rb
@@ -7,7 +7,7 @@ module Imap::Backup
   class Client::Default
     extend Forwardable
     def_delegators :imap, *%i(
-      append authenticate create expunge login
+      append authenticate create expunge login namespace
       responses uid_fetch uid_search uid_store
     )